requirements/requirements.txt
# imports
python-dotenv

# email
secure-smtplib

# cryptography
argon2-cffi
PyJWT

# api
fastapi
fastapi-mail
uvicorn[standard]

# database
sqlalchemy[asyncio]
asyncpg
psycopg2
alembic
aiosqlite
